+++ analyzer-models/src/main/java/ffb/analyzer/models/espn/TeamScore.java
@@ -12,10 +12,10 @@ public class TeamScore extends EspnEntity<TeamScore> {
     private int teamId;
 
     @JsonProperty("totalPoints")
-    private float points;
+    private double points;
 
     @JsonProperty("pointsByScoringPeriod")
-    private Map<Integer, Float> scores;
+    private Map<Integer, Double> scores;
 
     public CumulativeScore getCumulativeScore() {
         return cumulativeScore;
@@ -41,19 +41,19 @@ public class TeamScore extends EspnEntity<TeamScore> {
         this.teamId = teamId;
     }
 
-    public float getPoints() {
+    public double getPoints() {
         return points;
     }
 
-    public void setPoints(float points) {
+    public void setPoints(double points) {
         this.points = points;
     }
 
-    public Map<Integer, Float> getScores() {
+    public Map<Integer, Double> getScores() {
         return scores;
     }
 
-    public void setScores(Map<Integer, Float> scores) {
+    public void setScores(Map<Integer, Double> scores) {
         this.scores = scores;
     }
 }
